Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/body-onerror-compile-error-expected.txt

CONSOLE MESSAGE: line 27: SyntaxError: Unexpected token ')'

FAIL <body onerror> - compile error in <script> assert_equals: first arg expected "string" but got "object"
FAIL <body onerror> - compile error in <script> (column) assert_equals: fourth arg expected "number" but got "undefined"
